# swagger/test_spec.json
{
  "swagger": "2.0",
  "paths": {
...
    "/v1/vehicles/{vehicleId}": {
      "get": {
        "summary": "/v1/vehicles/{vehicleId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "Vehicle"
        ]
      },
      "patch": {
        "summary": "/v1/vehicles/{vehicleId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "Vehicle"
        ]
      },
      "delete": {
        "summary": "/v1/vehicles/{vehicleId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "Vehicle"
        ]
      }
    },
...
    "/v1/map-points/{mapPointId}": {
      "get": {
        "summary": "/v1/map-points/{mapPointId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mapPointId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "MapPoint"
        ]
      },
      "delete": {
        "summary": "/v1/map-points/{mapPointId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mapPointId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "MapPoint"
        ]
      }
    }
  },
...
  "info": {
    "title": "mytesla-api",
    "version": "0.1.0",
    "license": {
      "name": "MIT"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nAPI for myTesla"
  }
}